+++ src/components/story_container.js
@@ -23,39 +23,9 @@ class StoryContainer extends React.Component {
           focusedStory={this.props.focusedStory}
           lastStory={this.props.stories.length - 1}
           setFocus={this.props.setFocus} />
-        <h3>
-          Owner: {this.props.story.owner}
-        </h3>
-        <div>
-          {this.renderImages()}
-        </div>
-        {this.renderVideo()}
+        <StoryContent story={this.props.story} />
       </div>
     );
-  }
-
-  renderImages () {
-    var images = [];
-    this.props.story.images.forEach((image, index) => (
-      images.push (
-        <img
-          width="300px"
-          src={image}
-          key={index} />
-      )));
-    return images;
-  }
-
-  renderVideo () {
-    if (this.props.story.video) {
-      return (
-        <a
-          href={this.props.story.video}
-          target="_blank" >
-          A video about the struggle at this address.
-        </a>
-      );
-    }
   }
 }
